requirements.txt
streamlit==1.39.0
mysql-connector-python==9.0.0
pandas==2.2.3
matplotlib==3.9.2
datetime
python-dotenv==1.0.1
